APIs/graphql/requirements.txt
fastapi
uvicorn
httpie
strawberry-graphql
sqlalchemy
psycopg2
python-dotenv